fix: move BarrelBlockEntityAccessor
